# api/openapi/v1.yaml
...
openapi: "3.1.0"
info:
  title: GreenMetrics API
  version: "1.0.0"
  description: |
    Energy monitoring + carbon accounting API for Italian energy-intensive SMEs.
    Covers CSRD/ESRS E1, Piano Transizione 5.0 attestazione, Conto Termico 2.0,
    Certificati Bianchi (TEE), D.Lgs. 102/2014 audit energetico.

    All errors are RFC 7807 Problem Details. All timestamps are RFC 3339 UTC.
    All `tenant_id` and `id` fields are UUIDv4. All monetary values are
    `(amount_cents int64, currency ISO-4217 string)` — never float.
  contact:
    name: GreenMetrics Support
    email: support@greenmetrics.it
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8082
    description: Local dev (docker compose)
  - url: https://staging.greenmetrics.it
    description: Staging (eu-south-1)
  - url: https://app.greenmetrics.it
    description: Production (eu-south-1)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256-signed JWT with `kid` claim. Bearer prefix required.
        Access token TTL: 15 min default. Refresh token TTL: 720 h default.
        Sentinel secret rejected at boot in production.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        RFC 9457 / Stripe-style idempotency key. Required on POST in production.
        Stored 24h with response replay; conflict on different request hash returns 422.
      schema:
        type: string
        format: uuid
    PageLimit:
      name: limit
      in: query
      description: Maximum number of items per page (default 50, max 500).
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    PageOffset:
      name: offset
      in: query
      description: Page offset.
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference identifying the problem type.
          example: "https://greenmetrics.it/problems/validation-error"
        title:
          type: string
          description: Short, human-readable summary.
          example: "Validation error"
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: HTTP status code.
          example: 422
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying the specific occurrence.
        code:
          type: string
          description: Machine-readable error code (project-specific).
          example: "EMAIL_INVALID"
...
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp in UTC with offset (Z).
      example: "2026-04-25T12:34:56Z"
    TenantID:
      type: string
      format: uuid
      description: UUIDv4 tenant identifier; stable across the portfolio.
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    Meter:
      type: object
      required: [id, label, meter_type, protocol, unit]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          $ref: "#/components/schemas/TenantID"
        label:
          type: string
        meter_type:
          type: string
          enum: [electricity, gas, water, thermal, photovoltaic, ev_charger, pulse]
        protocol:
          type: string
          enum: [modbus_tcp, modbus_rtu, mbus, sunspec, ocpp, pulse, spd, manual]
        unit:
          type: string
        site:
          type: string
        cost_centre:
          type: string
        serial_no:
          type: string
        pod_code:
          type: string
          pattern: '^IT001E[0-9A-Z]{8}$'
          description: Italian POD electricity supply code.
        pdr_code:
          type: string
          description: Italian PDR gas supply code (14 digits).
        endpoint:
          type: string
        slave_addr:
          type: integer
        active:
          type: boolean
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
...
  responses:
    Problem400:
      description: Bad Request (validation failure, malformed body).
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Problem401:
      description: Unauthorized (missing or invalid token).
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Problem403:
      description: Forbidden (RBAC permission denied).
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    Problem422:
      description: Unprocessable Entity (idempotency conflict, semantic validation).
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
paths:
...
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      operationId: refresh
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: New access token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  expires_in: { type: integer }
                  token_type: { type: string, enum: [Bearer] }
        "401": { $ref: "#/components/responses/Problem401" }
...
  /api/v1/meters:
    get:
      summary: List meters (per-tenant)
      operationId: listMeters
      tags: [Meters]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageOffset"
      responses:
        "200":
          description: Meter list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Meter" }
        "401": { $ref: "#/components/responses/Problem401" }
    post:
      summary: Create meter
      operationId: createMeter
      tags: [Meters]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Meter" }
      responses:
        "201":
          description: Meter created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meter" }
        "400": { $ref: "#/components/responses/Problem400" }
        "401": { $ref: "#/components/responses/Problem401" }
        "403": { $ref: "#/components/responses/Problem403" }
        "422": { $ref: "#/components/responses/Problem422" }
...